Refactor to Feature-Based Architecture
## 変更内容

### アーキテクチャの改善
- レイヤーベース（Controllers/, Views/, Services/）から機能ベース（Features/）に変更
- 1つの機能に関連するすべてのファイル（Controller、View、Service）を同じフォルダにまとめる

### 新しいフォルダ構造
```
Features/
  ├── Home/
  │   ├── HomeController.cs
  │   └── Views/Index.cshtml
  ├── Calculator/
  │   ├── CalculatorController.cs
  │   ├── CalculatorService.cs
  │   └── Views/Index.cshtml
  ├── Orders/
  │   ├── OrdersController.cs
  │   ├── OrderService.cs
  │   ├── PricingService.cs
  │   └── Views/Index.cshtml
  └── Supabase/
      ├── SupabaseService.cs
      └── ISupabaseService.cs
```

### ファイル移動
- `Controllers/` → `Features/[Feature]/`
- `Views/[Feature]/` → `Features/[Feature]/Views/`
- Serviceは既に正しい場所に配置済み

### Namespace更新
- `BlazorApp.Controllers` → `BlazorApp.Features.[Feature]`
- Views/_ViewImports.cshtml を更新して新しいnamespaceを参照

### 利点
1. **保守性向上**: 1つの機能を変更する際、1つのフォルダ内で完結
2. **可読性向上**: 機能ごとに関連ファイルがまとまっている
3. **スケーラビリティ**: 新機能追加時にフォルダを作るだけ
4. **チーム開発**: 機能ごとに担当を分けやすい

### テスト結果
- ビルド: 成功
- 単体テスト: 27/27 合格
