+++ src/BlazorApp/Features/Home/Views/Index.cshtml
