+++ src/BlazorApp/Views/_ViewImports.cshtml
@@ -1,4 +1,6 @@
 @using BlazorApp
-@using BlazorApp.Controllers
+@using BlazorApp.Features.Home
+@using BlazorApp.Features.Calculator
+@using BlazorApp.Features.Orders
 @using BlazorApp.Services
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
